✨ [Feat] 불필요한 임시테스트 코드 제거(더미) 및 조회 로직 서비스/레포/컨트롤러 작성

+++ src/main/java/com/onenth/OneNth/domain/product/repository/PurchaseItemRepositoryCustom.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 // 수동 쿼리메서드 정의(구현은 Impl)
 public interface PurchaseItemRepositoryCustom {
-    List<PurchaseItemListDTO> findByRegion(Long regionId);
+    List<PurchaseItemListDTO> findByRegion(Integer regionId);
     List<PurchaseItemListDTO> findByCategoryAndRegions(String category, List<Long> regionIds);
 }
